Add check if claim address fits signer account
The merkle-drop flow has been extended. The claim button gets disabled,
if the default signer account (of the Web3 enabled browser or MetaMask
plugin) does not match the address that has been inserted to check
eligibility and get the proof.
Similar to the chain state implementation, also the default account gets
fetched in an interval. Based on both state variables of the address and
the account, there is an additional state which defines if the user can
claim his tokens by the equality of both values.

Related to: #107