Fix merkle drop claim flow for missing permission
The test if the address to claim the token for matches the default
account of the wallet does not work if the user has not granted
permission for this website. In such a case the workflow would lock the
user, since the permission access happens after he clicks the button to
claim.
Therefore the flow has been fixed to show the message and disable the
button only if the permissions are already granted. Additionally the
check will happen again, just before signing the transaction.

Please mind that many of the changes in this commit come from
restructuring the claim function. To improve its readability, it has
been changed from a `.then().catch()` to async-await.

Related to: trustlines-protocol/trustlines-foundation-website#107
See also: #2