Adjust merkle drop claim message font color
The displayed checkmark when an address is eligible was invisible,
because the font color was the same as the background. It has become
applied the same color as the header next to it.
By the input of our UX designer Sash, the color for both element has
been changed to bright white. The light gray did just not provide enough
contrast to make it well readable.

Related to: trustlines-network/project#758