Add T&C modal for manual merkle claim
Add a second T&C modal instance which gets opened when the user want to
manually claim his tokens. It behaves differently than the already
existing one. While this does trigger the token claim, the new one
simply shows the arguments for a manual claim.

Additionally both flows to accept the T&C a state if the user has
already accepted them. So if the user must accept them only once, and it
applies for both, no matter in which order he operates (if at all).

Related to: trustlines-network/project#776