Add feature for newsletter subscription
The user should be able to register himself to the mailing list which is
managed on Mailchimp. Therefore does each subpage has a new section
above the footer which allows to insert a mail address that will be
added.
Additionally does a banner in the hero section of the landing page offer
the same option. This banner is closable. In contrast to the other
sections which are static does the banner disappear for ever if the
user closed it or subscribed. Last option does also work if the user
subscribed via bottom section. The local storage is used to remember
that.

The user must accept the privacy policy, GDPR and enter a valid mail
address to enable the subscribe button.

As interface to Mailchimp their suggested form action link is used. This
will forward the user to Mailchimp itself if they click on subscribe. In
most cases this will simply show a success message and let the user
return. In case any error occurred (like subscribe twice) it gets
handled by Mailchimp and not this website. It is not possible to use
the Mailchimp API from the client side, since their servers do not allow
CORS requests (as also their documentation states).
The configuration file do include the form link. The dev version refers
to a test mail-list which can be abused for tests.

Webpack started to optimize the package generation due to the shared
React dependency with the merkle drop. This required to load the new
generated common.js package.

Related to: #30 #31
See also: trustlines-network/project#797