Clean up withdraw amount retrieval
The implementation to retrieve the withdrawn amount by its event was
not clean. I was using an argument to filter events by an transaction
hash that is not documented. Even if this was accepted and works in
the testing environment this must not work in production. Therefore this
has been removed for a straight API usage of web3js and do the filtering
for the users transaction manually.

Related to: #17
See also: #19